extract collection reading logic to its own service
