+++ src/HackF5.UnitySpy.HearthstoneLib/MindVision.cs
@@ -6,6 +6,7 @@ namespace HackF5.UnitySpy.HearthstoneLib
     using System.Diagnostics;
     using System.Collections.Generic;
     using System.Linq;
+    using HackF5.UnitySpy.HearthstoneLib.Collection;
 
     public class MindVision
     {
@@ -19,41 +20,9 @@ namespace HackF5.UnitySpy.HearthstoneLib
 
         public List<CollectionCard> GetCollection()
         {
-            List<CollectionCard> collectionCards = new List<CollectionCard>();
-            var collectibleCards = image["CollectionManager"]["s_instance"]?["m_collectibleCards"];
-            if (collectibleCards != null)
-            {
-                var items = collectibleCards["_items"];
-                int size = collectibleCards["_size"];
-                for (var i = 0; i < size; i++)
-                {
-                    string cardId = items[i]["m_EntityDef"]["m_cardIdInternal"];
-                    if (string.IsNullOrEmpty(cardId))
-                    {
-                        continue;
-                    }
-                    int count = items[i]["<OwnedCount>k__BackingField"];
-                    int premium = items[i]["m_PremiumType"];
-                    var card = collectionCards.Where(existingCard => existingCard.CardId == cardId).FirstOrDefault();
-                    if (card == null)
-                    {
-                        card = new CollectionCard
-                        {
-                            CardId = cardId
-                        };
-                        collectionCards.Add(card);
-                    }
-                    if (premium == 1)
-                    {
-                        card.PremiumCount = count;
-                    } 
-                    else
-                    {
-                        card.Count = count;
-                    }
-                }
-            }
-            return collectionCards;
+            return new CollectionReader(image).GetCollection();
         }
+
+
     }
 }
